rm deprecated class
