+++ brief05/.env
@@ -1,3 +1,3 @@
-DB_DSN=mysql:host=localhost;port=3306;dbname=honeyside
+DB_DSN=mysql:host=localhost;port=3307;dbname=honeyside
 DB_USER=root
 DB_PWD=debuneko/77
